openmetadata/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openmetadata-ingestion[mysql]==1.6.8",
    "pyyaml>=6.0.2",
]
